+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 **2.1.10 - 06/09/25**
 
   - Type-hinting: Fix mypy errors in logs.py
-  
+  - Type-hinting: Fix mypy errors in cli_tools.py
+
 **2.1.9 - 05/30/25**
 
   - Type-hinting: Fix mypy errors in conftest.py